use existing topicToTag instead of topicToLogo
