+++ src/data/homepage.ts
@@ -1,5 +1,5 @@
 import { EThirdPartyLogoVariant } from '@mdb/flora';
-import { productToLogo } from '../utils/logo';
+import { productToLogo } from '../utils/product-to-logo';
 
 export const cardsLanguagesData = [
     {
