+++ src/utils/language-to-logo.ts
@@ -1,0 +1,24 @@
+import { EThirdPartyLogoVariant } from '@mdb/flora';
+
+export const languageToLogo: { [key: string]: EThirdPartyLogoVariant } = {
+    BASH: EThirdPartyLogoVariant.BASH,
+    Bash: EThirdPartyLogoVariant.BASH,
+    C: EThirdPartyLogoVariant.C,
+    'C++': EThirdPartyLogoVariant.C_PLUS_PLUS,
+    'C#': EThirdPartyLogoVariant.C_SHARP,
+    CSharp: EThirdPartyLogoVariant.C_SHARP,
+    Dart: EThirdPartyLogoVariant.DART,
+    Go: EThirdPartyLogoVariant.GO,
+    Java: EThirdPartyLogoVariant.JAVA_LOGOMARK,
+    JavaScript: EThirdPartyLogoVariant.JAVASCRIPT,
+    Kotlin: EThirdPartyLogoVariant.KOTLIN,
+    PHP: EThirdPartyLogoVariant.PHP,
+    Python: EThirdPartyLogoVariant.PYTHON_LOGOMARK,
+    Ruby: EThirdPartyLogoVariant.RUBY_LOGOMARK,
+    Rust: EThirdPartyLogoVariant.RUST,
+    Scala: EThirdPartyLogoVariant.SCALA,
+    Swift: EThirdPartyLogoVariant.SWIFT,
+    TypeScript: EThirdPartyLogoVariant.TYPESCRIPT,
+    Typescript: EThirdPartyLogoVariant.TYPESCRIPT,
+    Kafka: EThirdPartyLogoVariant.KAFKA,
+};
