+++ src/utils/technology-to-logo.ts
@@ -1,0 +1,49 @@
+import { EThirdPartyLogoVariant } from '@mdb/flora';
+
+export const technologyToLogo: { [key: string]: EThirdPartyLogoVariant } = {
+    // Missing: Docker, FastAPI, Kubernetes, Heroku, NextJS, NodeJS, Unity
+    Android: EThirdPartyLogoVariant.ANDROID,
+    AWS: EThirdPartyLogoVariant.AWS,
+    Azure: EThirdPartyLogoVariant.AZURE,
+    Cloudflare: EThirdPartyLogoVariant.CLOUDFLARE,
+    Django: EThirdPartyLogoVariant.DJANGO,
+    Docker: EThirdPartyLogoVariant.DOCKER,
+    '.Net Framework': EThirdPartyLogoVariant.DOTNET, // Incorrect
+    '.NET': EThirdPartyLogoVariant.DOTNET,
+    Excel: EThirdPartyLogoVariant.EXCEL,
+    FastAPI: EThirdPartyLogoVariant.FASTAPI,
+    FastApi: EThirdPartyLogoVariant.FASTAPI, // Incorrect
+    Flask: EThirdPartyLogoVariant.FLASK,
+    Flutter: EThirdPartyLogoVariant.FLUTTER,
+    GCP: EThirdPartyLogoVariant.GOOGLE_CLOUD_LOGOMARK,
+    'GitHub Actions': EThirdPartyLogoVariant.GITHUB_ACTIONS,
+    'Github Actions': EThirdPartyLogoVariant.GITHUB_ACTIONS, // Incorrect
+    GraphQL: EThirdPartyLogoVariant.GRAPHQL,
+    iOS: EThirdPartyLogoVariant.IOS,
+    'Jetpack Compose': EThirdPartyLogoVariant.JETPACK_COMPOSE,
+    Kafka: EThirdPartyLogoVariant.KAFKA,
+    Kubernetes: EThirdPartyLogoVariant.KUBERNETES,
+    Netlify: EThirdPartyLogoVariant.NETLIFY,
+    Nextjs: EThirdPartyLogoVariant.NEXTJS, // Incorrect
+    'Next.js': EThirdPartyLogoVariant.NEXTJS,
+    Nodejs: EThirdPartyLogoVariant.NODEJS, // Incorrect
+    'Node.js': EThirdPartyLogoVariant.NODEJS,
+    Pandas: EThirdPartyLogoVariant.PANDAS,
+    Parquet: EThirdPartyLogoVariant.PARQUET,
+    'Postman API': EThirdPartyLogoVariant.POSTMAN,
+    Prisma: EThirdPartyLogoVariant.PRISMA,
+    RaspberryPi: EThirdPartyLogoVariant.RASPBERRY_PI, // Incorrect
+    'Raspberry Pi': EThirdPartyLogoVariant.RASPBERRY_PI,
+    React: EThirdPartyLogoVariant.REACT,
+    'React Native': EThirdPartyLogoVariant.REACT,
+    SQL: EThirdPartyLogoVariant.SQL,
+    Spark: EThirdPartyLogoVariant.SPARK,
+    Spring: EThirdPartyLogoVariant.SPRING,
+    Svelte: EThirdPartyLogoVariant.SVELTE,
+    Terraform: EThirdPartyLogoVariant.TERRAFORM,
+    TLS: EThirdPartyLogoVariant.TLS,
+    Unity: EThirdPartyLogoVariant.UNITY,
+    Vercel: EThirdPartyLogoVariant.VERCEL,
+    'VS Code': EThirdPartyLogoVariant.VS_CODE,
+    Xamarin: EThirdPartyLogoVariant.XAMARIN,
+};
